first non error version
