+++ pumpkin/src/plugin/api/context.rs
@@ -1,12 +1,9 @@
 use std::{fs, path::Path, sync::Arc};
 
-use pumpkin_event::{Event, EventPriority};
 use pumpkin_util::PermissionLvl;
-use tokio::sync::RwLock;
 
 use crate::{
     entity::player::Player,
-    plugin::{EventHandler, HandlerMap, TypedEventHandler},
     server::Server,
 };
 
@@ -15,19 +12,16 @@ use super::PluginMetadata;
 pub struct Context {
     metadata: PluginMetadata<'static>,
     pub server: Arc<Server>,
-    handlers: Arc<RwLock<HandlerMap>>,
 }
 impl Context {
     #[must_use]
     pub fn new(
         metadata: PluginMetadata<'static>,
         server: Arc<Server>,
-        handlers: Arc<RwLock<HandlerMap>>,
     ) -> Self {
         Self {
             metadata,
             server,
-            handlers,
         }
     }
 
@@ -51,28 +45,5 @@ impl Context {
     ) {
         let mut dispatcher_lock = self.server.command_dispatcher.write().await;
         dispatcher_lock.register(tree, permission);
-    }
-
-    pub async fn register_event<E: Event + 'static, H>(
-        &self,
-        handler: H,
-        priority: EventPriority,
-        blocking: bool,
-    ) where
-        H: EventHandler<E> + 'static,
-    {
-        let mut handlers = self.handlers.write().await;
-
-        let handlers_vec = handlers
-            .entry(E::get_name_static())
-            .or_insert_with(Vec::new);
-
-        let typed_handler = TypedEventHandler {
-            handler,
-            priority,
-            blocking,
-            _phantom: std::marker::PhantomData,
-        };
-        handlers_vec.push(Box::new(typed_handler));
     }
 }
